requirements.txt
pyspark==3.5.1
requests==2.31.0
backoff==2.2.1
jsonpointer==2.1
configparser==6.0.1
psycopg2-binary==2.9.9
